+++ release_notes.md
@@ -35,9 +35,12 @@ The Objective-C API has been updated and your code will break!
 * `columnTypeOfColumn:` renamed to `columnTypeOfColumnWithIndex` in `TDBTable`.
 * `columnNameOfColumn:` renamed to `nameOfColumnWithIndex:` in `TDBTable`.
 * `addColumnWithName:andType:` renamed to `addColumnWithName:type:` in `TDBDescriptor`.
+* readWithBlock: renamed to readTransactionUsingBlock: in `TDBContext`
+* writeWithBlock:error: renamed to writeTransactionUsingBlock:error: in `TDBContext`
 
 
 ### Enhancements:
+
 *